+++ assets/javascripts/kitten/components/images/image-cropper.test.js
@@ -166,15 +166,15 @@ describe('<ImageCropper />', () => {
       })
     })
   })
-  describe('with activateDelete prop at false', () => {
+  describe('with deletable prop at false', () => {
     const component = mount(
       <ImageCropper
-        activateDelete={ false }
+        deletable={ false }
         imageSrc="custom-file.jpg"
       />
     )
 
-    it('doesn\'t render the cancel button', () => {
+    it('doesn\'t render the delete button', () => {
       const deleteButton = component.find('.k-SimpleUploader__buttonIcon')
       expect(deleteButton).to.have.length(0)
     })