+++ assets/javascripts/kitten/components/uploaders/uploader.js
@@ -61,7 +61,7 @@ export class Uploader extends React.Component {
         onClick={ this.handleClick }
         onCancel={ this.handleCancel }
         disabled={ this.props.disabled }
-        activateDelete={ this.props.activateDelete }
+        deletable={ this.props.deletable }
       />
     )
   }